feat: Implement logging
