+++ main.py
@@ -1,12 +1,15 @@
-import argparse
+import logging
 
 from src.application.use_cases import DownloadAudioUseCase
 from src.infrastructure.file import EnglishAudioFileWriter
 from src.infrastructure.repositories import EnglishAudioRepository
 from src.presentation.cli import CommandHandler
+from src.utils import setup_argparse, setup_logger
 
 
 def main() -> None:
+    setup_logger()
+
     parser = setup_argparse()
     args = parser.parse_args()
 
@@ -17,26 +20,12 @@ def main() -> None:
         )
     )
 
-    # TODO: エラーハンドリングを書く
     command_handler.call_action(args)
 
 
-def setup_argparse() -> argparse.ArgumentParser:
-    parser = argparse.ArgumentParser('pyaudio command handler')
-    subparsers = parser.add_subparsers(
-        dest='command', required=True, help='Available commands'
-    )
-
-    download_parser = subparsers.add_parser(
-        'download', help='Download audio file'
-    )
-    download_parser.add_argument('url', type=str, help='URL of the audio file')
-    download_parser.add_argument(
-        'file_path', type=str, help='Path to save the downloaded audio file'
-    )
-
-    return parser
-
-
 if __name__ == '__main__':
-    main()
+    try:
+        main()
+    except Exception as e:
+        logging.error(e)
+        print('Application failed. Please check the logs.')
